# specs/005-create-other-pages/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: RentSight API - New Pages
  description: API endpoints for Settings, Properties, Help, Reports, and About pages
  version: 1.0.0
  contact:
    name: RentSight Development Team
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://rentsight.app/api
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /user/profile:
    get:
      tags: [User]
      summary: Get user profile
      description: Returns authenticated user's profile information
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    
    put:
      tags: [User]
      summary: Update user profile
      description: Updates authenticated user's profile information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdate'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
  /user/password:
    put:
      tags: [User]
      summary: Change user password
      description: Changes authenticated user's password (requires current password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - currentPassword
                - newPassword
              properties:
                currentPassword:
                  type: string
                  format: password
                newPassword:
                  type: string
                  format: password
                  minLength: 8
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Password updated successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Current password is incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
  /user/preferences:
    get:
      tags: [User]
      summary: Get user preferences
      description: Returns authenticated user's application preferences
      responses:
        '200':
          description: User preferences
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferences'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    
    put:
      tags: [User]
      summary: Update user preferences
      description: Updates authenticated user's application preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPreferencesUpdate'
      responses:
        '200':
          description: Preferences updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferences'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  
  # ========================================
  # Reports Endpoints
  # ========================================
  /reports/generate:
    post:
      tags: [Reports]
      summary: Generate report
      description: Generates a report based on specified type and filters
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportRequest'
      responses:
        '200':
          description: Report generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportData'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase JWT token from authentication
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserProfileUpdate:
      type: object
      properties:
        email:
          type: string
          format: email
    UserPreferences:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        currency_format:
          type: string
          example: USD
        date_format:
          type: string
          enum: [MM/DD/YYYY, DD/MM/YYYY, YYYY-MM-DD]
        language:
          type: string
          example: en
        default_view:
          type: string
          enum: [dashboard, properties, reports, settings]
        theme_preference:
          type: string
          enum: [light, dark, system]
          nullable: true
        preferences:
          type: object
          additionalProperties: true
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserPreferencesUpdate:
      type: object
      properties:
        currency_format:
          type: string
        date_format:
          type: string
          enum: [MM/DD/YYYY, DD/MM/YYYY, YYYY-MM-DD]
        language:
          type: string
        default_view:
          type: string
          enum: [dashboard, properties, reports, settings]
        theme_preference:
          type: string
          enum: [light, dark, system]
        preferences:
          type: object
          additionalProperties: true
    
    # Report Schemas
    ReportRequest:
      type: object
      required:
        - reportType
      properties:
        reportType:
          type: string
          enum: [income_summary, expense_breakdown, tax_report]
        filters:
          $ref: '#/components/schemas/ReportFilters'
    ReportFilters:
      type: object
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        propertyId:
          type: string
          format: uuid
        tagIds:
          type: array
          items:
            type: string
            format: uuid
        platform:
          type: string
        category:
          type: string
    ReportData:
      type: object
      properties:
        reportType:
          type: string
        generated_at:
          type: string
          format: date-time
        filters:
          $ref: '#/components/schemas/ReportFilters'
        data:
          type: object
          description: Report-specific data structure
          additionalProperties: true
        summary:
          type: object
          properties:
            totalIncome:
              type: number
            totalExpenses:
              type: number
            netIncome:
              type: number
            entryCount:
              type: integer
    
    # Help Schemas
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: User-friendly error message
        code:
          type: string
          description: Error code for programmatic handling
        details:
          type: object
          additionalProperties: true
          description: Additional error details
  responses:
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Validation failed
            code: VALIDATION_ERROR
            details:
              field: name
              message: Name is required
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            code: UNAUTHORIZED
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Internal server error
            code: INTERNAL_ERROR